Fix multiple image styles from squeezing or overflowing
